+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/InitiatorUserInfoProcessor.java
@@ -21,6 +21,7 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  */
 
 import com.walmartlabs.concord.project.InternalConstants;
+import com.walmartlabs.concord.server.process.pipelines.processors.signing.Signing;
 import com.walmartlabs.concord.server.security.ldap.LdapManager;
 
 import javax.inject.Inject;
@@ -30,7 +31,7 @@ import javax.inject.Named;
 public class InitiatorUserInfoProcessor extends UserInfoProcessor {
 
     @Inject
-    public InitiatorUserInfoProcessor(LdapManager ldapManager) {
-        super(InternalConstants.Request.INITIATOR_KEY, ldapManager);
+    public InitiatorUserInfoProcessor(LdapManager ldapManager, Signing signing) {
+        super(InternalConstants.Request.INITIATOR_KEY, ldapManager, signing);
     }
 }