+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/UserInfoProcessor.java
@@ -20,7 +20,9 @@ package com.walmartlabs.concord.server.process.pipelines.processors;
  * =====
  */
 
+import com.walmartlabs.concord.server.ConcordApplicationException;
 import com.walmartlabs.concord.server.process.Payload;
+import com.walmartlabs.concord.server.process.pipelines.processors.signing.Signing;
 import com.walmartlabs.concord.server.security.ldap.LdapManager;
 import com.walmartlabs.concord.server.security.ldap.UserInfo;
 import org.slf4j.Logger;
@@ -29,24 +31,31 @@ import org.slf4j.LoggerFactory;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * Collects and stores the current user's data.
+ */
 public abstract class UserInfoProcessor implements PayloadProcessor {
 
     private static final Logger log = LoggerFactory.getLogger(UserInfoProcessor.class);
 
-    private final LdapManager ldapManager;
-
     private final String key;
+    private final LdapManager ldapManager;
+    private final Signing signing;
 
-    public UserInfoProcessor(String key, LdapManager ldapManager) {
+    public UserInfoProcessor(String key, LdapManager ldapManager, Signing signing) {
         this.key = key;
         this.ldapManager = ldapManager;
+        this.signing = signing;
     }
 
     @Override
     public Payload process(Chain chain, Payload payload) {
-        // collect and store the initiator's data
-
         UserInfo info = ldapManager.getCurrentUserInfo();
+
+        if (signing.isEnabled()) {
+            info = sign(info);
+        }
+
         Map<String, UserInfo> m = new HashMap<>();
         m.put(key, info);
 
@@ -54,5 +63,36 @@ public abstract class UserInfoProcessor implements PayloadProcessor {
 
         log.info("process ['{}'] -> done", payload.getProcessKey());
         return chain.process(payload);
+    }
+
+    private UserInfo sign(UserInfo i) {
+        if (i == null || i.getUsername() == null) {
+            return i;
+        }
+
+        try {
+            String s = signing.sign(i.getUsername());
+            return new SignedUserInfo(i, s);
+        } catch (Exception e) {
+            throw new ConcordApplicationException("Error while singing process data: " + e.getMessage(), e);
+        }
+    }
+
+    public static class SignedUserInfo extends UserInfo {
+
+        private static final long serialVersionUID = 1L;
+
+        private final String usernameSignature;
+
+        public SignedUserInfo(UserInfo source,
+                              String usernameSignature) {
+
+            super(source.getUsername(), source.getDisplayName(), source.getGroups(), source.getAttributes());
+            this.usernameSignature = usernameSignature;
+        }
+
+        public String getUsernameSignature() {
+            return usernameSignature;
+        }
     }
 }